+++ src/players/__init__.py
@@ -1,0 +1,2 @@
+from .human_player import HumanPlayer
+from .primitive_ai import PrimitiveAI
