+++ src/resources/words.txt
