+++ src/services/__init__.py
@@ -1,0 +1,1 @@
+from .game import start_game
